fix bugs

+++ pages/mahasiswa/done-ujian/index.js
@@ -1,6 +1,6 @@
 import Image from "next/image";
 import Header from "../../../components/Header";
-import { useEffect, useState } from "react";
+import { useEffect, useState, useRef } from "react";
 import { ToastContainer, toast } from "react-toastify";
 import "react-toastify/dist/ReactToastify.css";
 import { useRouter } from "next/router";
@@ -13,34 +13,30 @@ import jwtDecode from "jwt-decode";
 import Cookies from "js-cookie";
 import { postHasilUjian } from "../../../services/mahasiswa";
 
-export default function DoneUjian({ mahasiswa }) {
+export default function DoneUjian({
+  mahasiswa,
+  jawabanMahasiswa,
+  detailJadwalUjian,
+  tokenUjian,
+}) {
   const router = useRouter();
   const dispatch = useDispatch();
 
-  const { dataJawabanMahasiswa } = useSelector(
-    (state) => state.jawabanMahasiswa
-  );
-  const { dataDetailJadwalUjian } = useSelector(
-    (state) => state.detailJadwalUjian
-  );
-  const { dataTokenUjian } = useSelector((state) => state.tokenUjian);
-  const { dataSoalUjian } = useSelector((state) => state.soalUjian);
   const { form } = useSelector((state) => state.formHasilUjian);
   const { dataHasilUjian } = useSelector((state) => state.hasilUjian);
 
   const [jumlahBenar, setJumlahBenar] = useState(0);
   const [nilai, setNilai] = useState(0);
+  const [tampilkanNilai, setTampilkanNilai] = useState(0);
+  const [soalUjian, setSoalUjian] = useState([]);
 
   // START: Hitung Jawaban Benar & Nilai
   let arrayJumlahBenar = [];
   let countNilai = 0;
 
-  if (
-    Object.keys(dataJawabanMahasiswa).length > 0 &&
-    dataSoalUjian.length > 0
-  ) {
-    dataJawabanMahasiswa?.data.forEach((jwbMhs) => {
-      dataSoalUjian.forEach((soal) => {
+  if (Object.keys(jawabanMahasiswa).length > 0 && soalUjian.length > 0) {
+    jawabanMahasiswa?.data.forEach((jwbMhs) => {
+      soalUjian.forEach((soal) => {
         if (
           jwbMhs?.idSoal === soal?._id &&
           jwbMhs?.jawaban === soal?.kunciJawaban?.message
@@ -57,14 +53,32 @@ export default function DoneUjian({ mahasiswa }) {
   // END: Hitung Jawaban Benar & Nilai
 
   const onSubmit = async () => {
-    const response = await postHasilUjian(form);
-    if (response?.data?.status === "error")
-      return toast.error(response?.data?.message);
-    toast.success("Hasil ujian berhasil disimpan!");
-    localStorage.clear();
-    Cookies.remove("dn");
-    router.push("/mahasiswa/jadwal-ujian");
+    if (localStorage.getItem("frm")) {
+      const newFormLocalStorage = JSON.parse(
+        Buffer.from(localStorage.getItem("frm"), "base64").toString("utf-8")
+      );
+      const response = await postHasilUjian(newFormLocalStorage);
+      if (response?.data?.status === "error") {
+        toast.error(response?.data?.message);
+      } else {
+        toast.success("Hasil ujian berhasil disimpan!");
+        localStorage.clear();
+        Cookies.remove("dn");
+        Cookies.remove("ct");
+        Cookies.remove("jm");
+        Cookies.remove("dju");
+        Cookies.remove("tu");
+        router.push("/mahasiswa/jadwal-ujian");
+      }
+    }
   };
+
+  if (form?.nilai !== 0 && Object.keys(form).length >= 7) {
+    localStorage.setItem(
+      "frm",
+      Buffer.from(JSON.stringify(form), "utf-8").toString("base64")
+    );
+  }
 
   useEffect(() => {
     dispatch(setDataHasilUjian(mahasiswa?._id));
@@ -77,30 +91,48 @@ export default function DoneUjian({ mahasiswa }) {
 
     if (countNilai > 0) {
       setNilai(countNilai);
+      localStorage.setItem(
+        "nli",
+        Buffer.from(countNilai.toString(), "utf-8").toString("base64")
+      );
     }
 
-    dispatch(setFormPostHasilUjian("jadwalUjian", dataDetailJadwalUjian?._id));
+    if (localStorage.getItem("nli")) {
+      setTampilkanNilai(
+        parseInt(
+          Buffer.from(localStorage.getItem("nli"), "base64").toString("utf-8")
+        )
+      );
+    }
+
+    if (localStorage.getItem("slu")) {
+      setSoalUjian(
+        JSON.parse(
+          Buffer.from(localStorage.getItem("slu"), "base64").toString("utf-8")
+        )
+      );
+    }
+
+    dispatch(setFormPostHasilUjian("jadwalUjian", detailJadwalUjian?._id));
     dispatch(setFormPostHasilUjian("mahasiswa", mahasiswa?._id));
     dispatch(
       setFormPostHasilUjian(
         "listJawaban",
-        JSON.stringify(dataJawabanMahasiswa?.data)
+        JSON.stringify(jawabanMahasiswa?.data)
       )
     );
     dispatch(setFormPostHasilUjian("jumlahBenar", jumlahBenar));
     dispatch(setFormPostHasilUjian("nilai", nilai));
-    dispatch(setFormPostHasilUjian("masuk", dataTokenUjian?.waktuMulai));
-    dispatch(
-      setFormPostHasilUjian("selesai", dataJawabanMahasiswa?.waktuSelesai)
-    );
+    dispatch(setFormPostHasilUjian("masuk", tokenUjian?.waktuMulai));
+    dispatch(setFormPostHasilUjian("selesai", jawabanMahasiswa?.waktuSelesai));
   }, [
     dataHasilUjian,
     router,
     dispatch,
-    dataDetailJadwalUjian?._id,
-    dataJawabanMahasiswa?.data,
-    dataJawabanMahasiswa?.waktuSelesai,
-    dataTokenUjian?.waktuMulai,
+    detailJadwalUjian?._id,
+    jawabanMahasiswa?.data,
+    jawabanMahasiswa?.waktuSelesai,
+    tokenUjian?.waktuMulai,
     mahasiswa?._id,
     arrayJumlahBenar.length,
     countNilai,
@@ -122,7 +154,7 @@ export default function DoneUjian({ mahasiswa }) {
           Selamat, ujian telah selesai! <br />
           <span className="text-base text-gray-400">
             Jangan lupa tekan tombol simpan untuk menyimpan hasil ujian. <br />
-            Nilai ujian Anda : {form?.nilai}
+            Nilai ujian Anda : {tampilkanNilai}
           </span>{" "}
         </h1>
         <button
@@ -172,9 +204,24 @@ export async function getServerSideProps({ req }) {
   const jwtToken = Buffer.from(tkn, "base64").toString("utf-8");
   const payload = jwtDecode(jwtToken);
   const mahasiswa = payload.mahasiswa;
+
+  const jm = req.cookies.jm;
+  const jawabanMahasiswa = JSON.parse(
+    Buffer.from(jm, "base64").toString("utf-8")
+  );
+
+  const dju = req.cookies.dju;
+  const detailJadwalUjian = JSON.parse(
+    Buffer.from(dju, "base64").toString("utf-8")
+  );
+  const tu = req.cookies.tu;
+  const tokenUjian = JSON.parse(Buffer.from(tu, "base64").toString("utf-8"));
   return {
     props: {
       mahasiswa,
+      jawabanMahasiswa,
+      detailJadwalUjian,
+      tokenUjian,
     },
   };
 }
